e0ed3ff0bc746633f377baf38b545eed87e847f9 Add check to prevent invalid sql when nothing to delte

+++ sites/default/civicrm/extensions/org.wikimedia.forgetme/api/v3/Logging/Forgetme.php
@@ -61,7 +61,9 @@ function civicrm_api3_logging_forgetme($params) {
     );
 
     $activitiesToDelete = array_diff($activities, explode(',', $activitiesToKeep));
-    CRM_Core_DAO::executeQuery("DELETE FROM log_civicrm_activity WHERE id IN (" . implode(',', $activitiesToDelete) . ")");
+    if (!empty($activitiesToDelete)) {
+      CRM_Core_DAO::executeQuery("DELETE FROM log_civicrm_activity WHERE id IN (" . implode(',', $activitiesToDelete) . ")");
+    }
   }
 
   $fieldsToForget = CRM_Forgetme_Metadata::getMetadataForEntity('Contact', 'forget_fields');
